+++ src/data/action.ts
@@ -1,11 +1,17 @@
 import { $TRACK, createMemo, createSignal, JSX, onCleanup, getOwner } from "solid-js";
 import { isServer } from "solid-js/web";
 import { useRouter } from "../routing.js";
-import type { RouterContext, Submission, SubmissionStub, Navigator, NarrowResponse } from "../types.js";
+import type {
+  RouterContext,
+  Submission,
+  SubmissionStub,
+  Navigator,
+  NarrowResponse
+} from "../types.js";
 import { mockBase } from "../utils.js";
-import { cacheKeyOp, hashKey, revalidate, cache } from "./cache.js";
+import { cacheKeyOp, hashKey, revalidate, query } from "./query.js";
 
-export type Action<T extends Array<any>, U> = (T extends [FormData] | []
+export type Action<T extends Array<any>, U, V = T> = (T extends [FormData] | []
   ? JSX.SerializableAttributeValue
   : unknown) &
   ((...vars: T) => Promise<NarrowResponse<U>>) & {
@@ -13,52 +19,64 @@ export type Action<T extends Array<any>, U> = (T extends [FormData] | []
     with<A extends any[], B extends any[]>(
       this: (this: any, ...args: [...A, ...B]) => Promise<NarrowResponse<U>>,
       ...args: A
-    ): Action<B, U>;
+    ): Action<B, U, V>;
   };
 
 export const actions = /* #__PURE__ */ new Map<string, Action<any, any>>();
 
-export function useSubmissions<T extends Array<any>, U>(
-  fn: Action<T, U>,
-  filter?: (arg: T) => boolean
+export function useSubmissions<T extends Array<any>, U, V>(
+  fn: Action<T, U, V>,
+  filter?: (input: V) => boolean
 ): Submission<T, NarrowResponse<U>>[] & { pending: boolean } {
   const router = useRouter();
   const subs = createMemo(() =>
-    router.submissions[0]().filter(s => s.url === fn.toString() && (!filter || filter(s.input)))
+    router.submissions[0]().filter(s => s.url === (fn as any).base && (!filter || filter(s.input)))
   );
   return new Proxy<Submission<any, any>[] & { pending: boolean }>([] as any, {
     get(_, property) {
       if (property === $TRACK) return subs();
       if (property === "pending") return subs().some(sub => !sub.result);
       return subs()[property as any];
+    },
+    has(_, property) {
+      return property in subs();
     }
   });
 }
 
-export function useSubmission<T extends Array<any>, U>(
-  fn: Action<T, U>,
-  filter?: (arg: T) => boolean
+export function useSubmission<T extends Array<any>, U, V>(
+  fn: Action<T, U, V>,
+  filter?: (input: V) => boolean
 ): Submission<T, NarrowResponse<U>> | SubmissionStub {
   const submissions = useSubmissions(fn, filter);
   return new Proxy(
     {},
     {
       get(_, property) {
-        if (submissions.length === 0 && property === "clear" || property === "retry") return (() => {});
+        if ((submissions.length === 0 && property === "clear") || property === "retry")
+          return () => {};
         return submissions[submissions.length - 1]?.[property as keyof Submission<T, U>];
       }
     }
   ) as Submission<T, NarrowResponse<U>>;
 }
 
-export function useAction<T extends Array<any>, U>(action: Action<T, U>) {
+export function useAction<T extends Array<any>, U, V>(action: Action<T, U, V>) {
   const r = useRouter();
-  return (...args: Parameters<Action<T, U>>) => action.apply({ r }, args);
+  return (...args: Parameters<Action<T, U, V>>) => action.apply({ r }, args);
 }
 
 export function action<T extends Array<any>, U = void>(
   fn: (...args: T) => Promise<U>,
   name?: string
+): Action<T, U>;
+export function action<T extends Array<any>, U = void>(
+  fn: (...args: T) => Promise<U>,
+  options?: { name?: string; onComplete?: (s: Submission<T, U>) => boolean }
+): Action<T, U>;
+export function action<T extends Array<any>, U = void>(
+  fn: (...args: T) => Promise<U>,
+  options: string | { name?: string; onComplete?: (s: Submission<T, U>) => boolean } = {}
 ): Action<T, U> {
   function mutate(this: { r: RouterContext; f?: HTMLFormElement }, ...variables: T) {
     const router = this.r;
@@ -73,6 +91,17 @@ export function action<T extends Array<any>, U = void>(
     function handler(error?: boolean) {
       return async (res: any) => {
         const result = await handleResponse(res, error, router.navigatorFactory());
+        let retry = null;
+        !o.onComplete?.({
+          ...submission,
+          result: result?.data,
+          error: result?.error,
+          pending: false,
+          retry() {
+            return retry = submission.retry();
+          }
+        });
+        if (retry) return retry;
         if (!result) return submission.clear();
         setResult(result);
         if (result.error && !form) throw result.error;
@@ -94,7 +123,7 @@ export function action<T extends Array<any>, U = void>(
           return !result();
         },
         clear() {
-          router.submissions[1](v => v.filter(i => i.input !== variables));
+          router.submissions[1](v => v.filter(i => i !== submission));
         },
         retry() {
           setResult(undefined);
@@ -105,15 +134,16 @@ export function action<T extends Array<any>, U = void>(
     ]);
     return p.then(handler(), handler(true));
   }
-
+  const o = typeof options === "string" ? { name: options } : options;
   const url: string =
     (fn as any).url ||
-    (name && `https://action/${name}`) ||
+    (o.name && `https://action/${o.name}`) ||
     (!isServer ? `https://action/${hashString(fn.toString())}` : "");
+  mutate.base = url;
   return toAction(mutate, url);
 }
 
-function toAction<T extends Array<any>, U>(fn: Function, url: string): Action<T, U> {
+function toAction<T extends Array<any>, U, V = T>(fn: Function, url: string): Action<T, U, V> {
   fn.toString = () => {
     if (!url) throw new Error("Client Actions need explicit names if server rendered");
     return url;
@@ -125,19 +155,20 @@ function toAction<T extends Array<any>, U>(fn: Function, url: string): Action<T,
     const newFn = function (this: RouterContext, ...passedArgs: B): U {
       return fn.call(this, ...args, ...passedArgs);
     };
+    newFn.base = (fn as any).base;
     const uri = new URL(url, mockBase);
     uri.searchParams.set("args", hashKey(args));
-    return toAction<B, U>(
+    return toAction<B, U, V>(
       newFn as any,
       (uri.origin === "https://action" ? uri.origin : "") + uri.pathname + uri.search
     );
   };
   (fn as any).url = url;
   if (!isServer) {
-    actions.set(url, fn as Action<T, U>);
+    actions.set(url, fn as Action<T, U, V>);
     getOwner() && onCleanup(() => actions.delete(url));
   }
-  return fn as Action<T, U>;
+  return fn as Action<T, U, V>;
 }
 
 const hashString = (s: string) =>
@@ -172,7 +203,7 @@ async function handleResponse(response: unknown, error: boolean | undefined, nav
   // invalidate
   cacheKeyOp(keys, entry => (entry[0] = 0));
   // set cache
-  flightKeys && flightKeys.forEach(k => cache.set(k, custom[k]));
+  flightKeys && flightKeys.forEach(k => query.set(k, custom[k]));
   // trigger revalidation
   await revalidate(keys, false);
   return data != null ? { data } : undefined;